Create switchcase.java
In this code, we will learn about switch case, Break, and default case. Switch cases are used to define the different statements inside the code where we also use break statements to get out of the case and switch to another case. default case is used as a default setting in the code which gives the output when the input doesn't match the program criteria.