Create practice3.java
In this code, the question has mentioned to calculate the income tax by taking the input from the user. here I have used only integers but we can use large numbers too. we have used if-else ladders to make the code look good and easy to read. Do check it !!