Create rockpaper.java
This program is written to design a basic game called Rock Paper Scissor which will be played between the user and the computer. The user input will be taken with the help of the scanner class and the computer will take Random numbers in between 1 to 3 with the help of the Random class. We have used if-else statements to apply the condition of the game. Do try to play this game !!